+++ dbm-ui/frontend/src/common/const/ticketTypes.ts
@@ -16,7 +16,7 @@ export enum TicketTypes {
   MYSQL_DELETE_CLEAR_DB = 'MYSQL_DELETE_CLEAR_DB',
   MYSQL_DUMP_DATA = 'MYSQL_DUMP_DATA', // mysql 数据导出
   MYSQL_EXCEL_AUTHORIZE_RULES = 'MYSQL_EXCEL_AUTHORIZE_RULES',
-  MYSQL_FIXPOINT = 'MYSQL_FIXPOINT', // mysql 构造
+  MYSQL_FIXPOINT_EXIST_CLUSTER = 'MYSQL_FIXPOINT_EXIST_CLUSTER', // mysql 构造到已有集群
   MYSQL_FLASHBACK = 'MYSQL_FLASHBACK',
   MYSQL_FORCE_IMPORT_SQLFILE = 'MYSQL_FORCE_IMPORT_SQLFILE',
   MYSQL_HA_APPLY = 'MYSQL_HA_APPLY',
@@ -49,6 +49,7 @@ export enum TicketTypes {
   MYSQL_ROLLBACK = 'MYSQL_ROLLBACK', // mysql 回档
   MYSQL_ROLLBACK_CLUSTER = 'MYSQL_ROLLBACK_CLUSTER',
   MYSQL_SEMANTIC_CHECK = 'MYSQL_SEMANTIC_CHECK', // 模拟执行
+  MYSQL_FIXPOINT_NEW_CLUSTER = 'MYSQL_FIXPOINT_NEW_CLUSTER', // mysql 构造到新集群
   MYSQL_SINGLE_APPLY = 'MYSQL_SINGLE_APPLY',
   MYSQL_SINGLE_DESTROY = 'MYSQL_SINGLE_DESTROY',
   MYSQL_SINGLE_DISABLE = 'MYSQL_SINGLE_DISABLE',
