+++ dbm-ui/frontend/src/locales/zh-cn.json
@@ -5101,5 +5101,6 @@
   "people于date修改了跟进内容": "{0} 于 {1} 修改了跟进内容",
   "通过全备 + binlog 的方式，将数据库恢复到过去的某一时间点或者某个指定备份文件的状态。数据可以构造到新临时单节点，可以选择已有的集群": "通过全备 + binlog 的方式，将数据库恢复到过去的某一时间点或者某个指定备份文件的状态。数据可以构造到新临时单节点，可以选择已有的集群",
   "在选择原集群上进行原地数据回滚，支持构造回档、库表闪回、记录级闪回": "在选择原集群上进行原地数据回滚，支持构造回档、库表闪回、记录级闪回",
+  "构造": "构造",
   "这行勿动！新增翻译请在上一行添加！": ""
 }
