+++ dbm-ui/frontend/src/views/db-manage/mysql/MYSQL_FIXPOINT_EXIST_CLUSTER/components/conflict-db-column/Index.vue
