+++ dbm-ui/frontend/src/views/ticket-center/common/ticket-detail/components/task-info/com-factory/mysql/FixpointExistCluster.vue
@@ -214,7 +214,7 @@
   type RowData = Props['ticketDetails']['details']['infos'][number];
 
   defineOptions({
-    name: TicketTypes.MYSQL_FIXPOINT,
+    name: TicketTypes.MYSQL_FIXPOINT_EXIST_CLUSTER,
     inheritAttrs: false,
   });
